lidhit/lidhit/backend/requirements.txt
asgiref==3.6.0
DateTime==4.9
Django==4.1.5
djangorestframework==3.14.0
phonenumbers==8.13.4
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
zope.interface==5.5.2
